--- game/test/mydungeon/DesktopLauncherTest.java
@@ -1,20 +1,0 @@
-package mydungeon;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.powermock.core.classloader.annotations.PrepareForTest;
-import org.powermock.modules.junit4.PowerMockRunner;
-import starter.DesktopLauncher;
-
-@RunWith(PowerMockRunner.class)
-@PrepareForTest({DesktopLauncher.class})
-public class DesktopLauncherTest {
-    /** A set-up method to initialize the tests. */
-    @Before
-    public void setUp() {}
-
-    /** A dummy test skeleton. Not implemented yet. */
-    @Test
-    public void testDesktopLauncherDummyTest() {}
-}
